[Feat] #32 - 공통되는 showAlert Extension으로 변경
